requirements.txt
streamlit 
python-dotenv
openai
pypdf
chromadb
typing